Refactoring the pdf file builders - no longer part of the 'claim file builder' and 'response file builder' - which used to be present when we generated lots of files from within the app
