+++ app/event_handlers/claim_pdf_file_handler.rb
@@ -1,7 +1,7 @@
 class ClaimPdfFileHandler
   def handle(claim)
     if claim.pdf_file.blank?
-      ClaimFileBuilder::BuildClaimPdfFile.new(claim, template_reference: claim.pdf_template_reference).call
+      BuildClaimPdfFileService.new(claim, template_reference: claim.pdf_template_reference).call
       claim.save!
     end
     Rails.application.event_service.publish('ClaimPdfFileAdded', claim)
