+++ app/event_handlers/response_pdf_file_handler.rb
@@ -1,6 +1,6 @@
 class ResponsePdfFileHandler
   def handle(response)
-    ResponseFileBuilder::BuildResponsePdfFile.call(response, template_reference: response.pdf_template_reference)
+    BuildResponsePdfFileService.call(response, template_reference: response.pdf_template_reference)
     response.save!
     Rails.application.event_service.publish('ResponsePdfFileAdded', response)
   end
